+++ commands/stop.js
@@ -1,11 +1,14 @@
 module.exports = {
 	name: 'stop',
 	description: 'Остановить воспроизведение аудио',
-	execute(DISCORDMessage) {
+	execute(DISCORDMessage, chack) {
 		if (!DISCORDMessage.member.voice.channel) return DISCORDMessage.channel.send('Вы должны быть в голосовом канале, чтобы остановить музыку!');
 		DISCORDMessage.client.queue.get(DISCORDMessage.guild.id).songs = [];
 		DISCORDMessage.client.queue.get(DISCORDMessage.guild.id).connection.dispatcher.end();
 		DISCORDMessage.client.queue.get(DISCORDMessage.guild.id).player.delete();
-		DISCORDMessage.delete();
+		if (chack) {
+			DISCORDMessage.delete();
+		}
+
 	},
 };
